# src/main/resources/swagger/sboot-planets.yaml
openapi: 3.0.1
info:
  title: Planet API
  description: Api.
  termsOfService: http://swagger.io/terms/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
  - url: http://localhost:8993
tags:
  - name: search
    description: Operations find planets
paths:
  /planets/v1/search:
    post:
      operationId: searchPlanets
      summary: "Search Planets"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PayloadSearchRepresentation'
      responses:
        200:
          description: Ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseRepresentation'
        400:
          description: Invalid Payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseRepresentation'
  /planets/v1:
    post:
      operationId: newPlanet
      summary: "New Planet"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanetRepresentation'
      responses:
        201:
          description: Recurso criado!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanetRepresentation'
        400:
          description: Invalid Payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseRepresentation'
  /planets/v1/{planetId}:
    delete:
      operationId: deletePlanet
      summary: "New Planet"
      parameters:
        - name: planetId
          in: path
          description: Delete Planet by ID
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: Recurso deletado!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanetRepresentation'
        400:
          description: Invalid Payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseRepresentation'
components:
  schemas:
    PayloadSearchRepresentation:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        page:
          type: integer
        totalElements:
          type: integer
        totalPages:
          type: integer
    ResponseRepresentation:
      type: object
      properties:
        page:
          type: integer
        totalPages:
          type: integer
        totalElements:
          type: integer
        elements:
          type: array
          items:
            $ref: '#/components/schemas/PlanetRepresentation'
    PlanetRepresentation:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        climate:
          type: string
        terrain:
          type: string
        films:
          type: integer
        createdAt:
          type: string
          format: date-time
        formattedDate:
          type: string